refactor: Application에 있는 playGame에 대한 역할을 Game으로 이동
