+++ src/main/kotlin/blackjack/controller/Application.kt
@@ -4,7 +4,6 @@ import blackjack.domain.Dealer
 import blackjack.domain.Game
 import blackjack.domain.GameCards
 import blackjack.domain.Names
-import blackjack.domain.Participant
 import blackjack.domain.Player
 import blackjack.view.InputView
 import blackjack.view.ResultView
@@ -23,7 +22,7 @@ fun main() {
 
     ResultView.printAllPlayerCards(game)
 
-    playGameWithParticipants(game)
+    game.playGameWithParticipants()
 
     ResultView.printAllResult(game)
 
@@ -34,36 +33,4 @@ fun main() {
     ResultView.printGameResults(dealer, game)
 }
 
-private fun playGameWithParticipants(game: Game) {
-    game.participants.forEach {
-        playGame(it, game)
-    }
 
-    drawIfSmallerThanMinimum(game.dealer, game)
-}
-
-private fun playGame(participant: Participant, game: Game) {
-    while (participant.isPlaying) {
-        val answer = InputView.askIfPlayerWantToMoreCard(participant.name)
-
-        drawOrStopByAnswer(answer, game, participant)
-
-        ResultView.printPlayerCards(participant.name, participant.cards)
-    }
-}
-
-private fun drawIfSmallerThanMinimum(dealer: Dealer, game: Game) {
-    if (dealer.isSmallerThanMinimumScore()) {
-        println("딜러는 16이하라 한장의 카드를 더 받았습니다.")
-        game.draw(dealer)
-    }
-}
-
-private fun drawOrStopByAnswer(answer: Boolean, game: Game, participant: Participant) {
-    if (answer) {
-        game.draw(participant)
-        return
-    }
-
-    participant.stop()
-}
